Rename index.js to definitions.js to avoid confusing multiple index files
